backend/requirements.txt
Flask
torch
torchvision
pillow
flask-cors